# Addon-docs, blocks: Prebundle dependencies