--- a/code/addons/docs/angular/index.js
+++ b/code/addons/docs/angular/index.js
@@ -1,7 +1,6 @@
 /* eslint-disable no-underscore-dangle */
-import { global } from '@storybook/global';
 
 export const setCompodocJson = (compodocJson) => {
   // @ts-expect-error (Converted from ts-ignore)
-  global.__STORYBOOK_COMPODOC_JSON__ = compodocJson;
+  globalThis.__STORYBOOK_COMPODOC_JSON__ = compodocJson;
 };