--- a/code/addons/docs/ember/index.js
+++ b/code/addons/docs/ember/index.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-underscore-dangle */
-import { global } from '@storybook/global';
 
 export const setJSONDoc = (jsondoc) => {
-  global.__EMBER_GENERATED_DOC_JSON__ = jsondoc;
+  globalThis.__EMBER_GENERATED_DOC_JSON__ = jsondoc;
 };