--- a/code/addons/docs/package.json
+++ b/code/addons/docs/package.json
@@ -97,6 +97,7 @@
     "prep": "jiti ../../../scripts/prepare/bundle.ts"
   },
   "dependencies": {
+    "@mdx-js/react": "^3.0.0",
     "@storybook/blocks": "workspace:*",
     "@storybook/csf-plugin": "workspace:*",
     "@storybook/react-dom-shim": "workspace:*",
@@ -106,7 +107,6 @@
   },
   "devDependencies": {
     "@mdx-js/mdx": "^3.0.0",
-    "@mdx-js/react": "^3.0.0",
     "@rollup/pluginutils": "^5.0.2",
     "@types/react": "^16.8.0 || ^17.0.0 || ^18.0.0",
     "react": "^18.2.0",