--- a/code/addons/docs/template/stories/toc/basic.stories.ts
+++ b/code/addons/docs/template/stories/toc/basic.stories.ts
@@ -1,6 +1,3 @@
-import { global as globalThis } from '@storybook/global';
-import { fn } from '@storybook/test';
-
 export default {
   component: globalThis.Components.Button,
   tags: ['autodocs'],