--- a/code/lib/blocks/package.json
+++ b/code/lib/blocks/package.json
@@ -44,7 +44,6 @@
   },
   "dependencies": {
     "@storybook/csf": "^0.1.11",
-    "@storybook/global": "^5.0.0",
     "@storybook/icons": "^1.2.10",
     "color-convert": "^2.0.1",
     "dequal": "^2.0.2",