--- a/code/lib/blocks/src/blocks/DocsContext.ts
+++ b/code/lib/blocks/src/blocks/DocsContext.ts
@@ -4,8 +4,6 @@ import { createContext } from 'react';
 
 import type { DocsContextProps, Renderer } from 'storybook/internal/types';
 
-import { global } from '@storybook/global';
-
 export type { DocsContextProps };
 
 // We add DocsContext to window. The reason is that in case DocsContext.ts is
@@ -13,11 +11,11 @@ export type { DocsContextProps };
 // we will have multiple DocsContext definitions - leading to lost context in
 // the React component tree.
 // This was specifically a problem with the Vite builder.
-if (global && global.__DOCS_CONTEXT__ === undefined) {
-  global.__DOCS_CONTEXT__ = createContext(null);
-  global.__DOCS_CONTEXT__.displayName = 'DocsContext';
+if (globalThis && globalThis.__DOCS_CONTEXT__ === undefined) {
+  globalThis.__DOCS_CONTEXT__ = createContext(null);
+  globalThis.__DOCS_CONTEXT__.displayName = 'DocsContext';
 }
 
-export const DocsContext: Context<DocsContextProps<Renderer>> = global
-  ? global.__DOCS_CONTEXT__
+export const DocsContext: Context<DocsContextProps<Renderer>> = globalThis
+  ? globalThis.__DOCS_CONTEXT__
   : createContext(null);