--- a/code/lib/blocks/src/components/IFrame.tsx
+++ b/code/lib/blocks/src/components/IFrame.tsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 
-import { global } from '@storybook/global';
 
-const { window: globalWindow } = global;
+const { window: globalWindow } = globalThis;
 
 interface IFrameProps {
   id: string;