--- a/code/lib/blocks/src/components/IFrame.tsx
+++ b/code/lib/blocks/src/components/IFrame.tsx
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 
-
 const { window: globalWindow } = globalThis;
 
 interface IFrameProps {