Works on https://github.com/storybookjs/storybook/issues/29038

## What I did

|  | **Before** | **After** |
|---|---|---|
| pkg-size | ![image](https://github.com/user-attachments/assets/972c14a5-12ba-454f-91c3-c32714fd81bb) | ![image](https://github.com/user-attachments/assets/ff3cedff-2973-41f1-8463-2304959bc61a) |
| npmgraph | ![image](https://github.com/user-attachments/assets/3794cb55-a8f4-464e-9c16-3ccfc61c74e4) | ![image](https://github.com/user-attachments/assets/c949b846-bdda-4e5e-a108-7b91636f90cf) |

### Sources

https://pkg-size.dev/@storybook%2Faddon-docs@8.4.0-alpha.5?no-peers

https://pkg-size.dev/@storybook%2Faddon-docs@0.0.0-pr-29301-sha-991d41fc?no-peers

https://npmgraph.js.org/?q=@storybook/addon-docs@8.4.0-alpha.5#hide=

https://npmgraph.js.org/?q=@storybook/addon-docs@0.0.0-pr-29301-sha-991d41fc#zoom=w

<!-- Briefly describe what your PR does -->

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:
- [x] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

     - `bug`: Internal changes that fixes incorrect behavior.
     - `maintenance`: User-facing maintenance tasks.
     - `dependencies`: Upgrading (sometimes downgrading) dependencies.
     - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
     - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
     - `documentation`: Documentation **only** changes. Will not show up in release changelog.
     - `feature request`: Introducing a new feature.
     - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
     - `other`: Changes that don't fit in the above categories.
   
   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-29301-sha-991d41fc`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-29301-sha-991d41fc sandbox` or in an existing project with `npx storybook@0.0.0-pr-29301-sha-991d41fc upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-29301-sha-991d41fc`](https://npmjs.com/package/storybook/v/0.0.0-pr-29301-sha-991d41fc) |
| **Triggered by** | @JReinhold |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`jeppe/prebundle-addon-docs`](https://github.com/storybookjs/storybook/tree/jeppe/prebundle-addon-docs) |
| **Commit** | [`991d41fc`](https://github.com/storybookjs/storybook/commit/991d41fc8a1bff88dc1a7e208446b234e503b8a9) |
| **Datetime** | Tue Oct  8 20:01:14 UTC 2024 (`1728417674`) |
| **Workflow run** | [11242809433](https://github.com/storybookjs/storybook/actions/runs/11242809433) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=29301`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  78.7 MB | 78.7 MB | 0 B | 0.83 | 0% |
| initSize |  148 MB | 147 MB | -1 MB | **-4.13** | -0.7% |
| diffSize |  69.3 MB | 68.3 MB | -1 MB | **-3.74** | -1.5% |
| buildSize |  6.79 MB | 6.79 MB | 4.49 kB | **5.4** | 0.1% |
| buildSbAddonsSize |  1.5 MB | 1.5 MB | -281 B | **-Infinity** | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.83 MB | 1.83 MB | 0 B | 0.58 | 0% |
| buildSbPreviewSize |  270 kB | 270 kB | 0 B | - | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.8 MB | 3.8 MB | -281 B | 0.29 | 0% |
| buildPreviewSize |  2.99 MB | 2.99 MB | 4.77 kB | **5.7** | 0.2% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  16.5s | 23.4s | 6.8s | **1.66** | 🔺29.3% |
| generateTime |  20.6s | 18.6s | -2s -1ms | -1.18 | -10.7% |
| initTime |  13.5s | 12.5s | -1s -27ms | **-1.37** | 🔰-8.2% |
| buildTime |  8.2s | 8.2s | 8ms | -0.8 | 0.1% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  6.7s | 5.8s | -870ms | -1.18 | -14.9% |
| devManagerResponsive |  4.1s | 3.9s | -251ms | -1.15 | -6.4% |
| devManagerHeaderVisible |  587ms | 541ms | -46ms | -0.78 | -8.5% |
| devManagerIndexVisible |  618ms | 563ms | -55ms | -0.87 | -9.8% |
| devStoryVisibleUncached |  1s | 851ms | -176ms | -0.91 | -20.7% |
| devStoryVisible |  617ms | 570ms | -47ms | -0.79 | -8.2% |
| devAutodocsVisible |  495ms | 494ms | -1ms | -0.31 | -0.2% |
| devMDXVisible |  519ms | 457ms | -62ms | -0.88 | -13.6% |
| buildManagerHeaderVisible |  492ms | 449ms | -43ms | -1.14 | -9.6% |
| buildManagerIndexVisible |  495ms | 460ms | -35ms | -1.15 | -7.6% |
| buildStoryVisible |  525ms | 716ms | 191ms | 1.09 | 26.7% |
| buildAutodocsVisible |  445ms | 504ms | 59ms | -0.05 | 11.7% |
| buildMDXVisible |  496ms | 437ms | -59ms | -0.84 | -13.5% |

<!-- BENCHMARK_SECTION -->

<!-- greptile_comment -->

## Greptile Summary

This PR focuses on prebundling dependencies for the addon-docs and blocks packages to reduce Storybook's install footprint. Key changes include:

- Removed several dependencies from `@storybook/blocks` package.json, moving them to devDependencies
- Replaced imports of `global` from '@storybook/global' with direct use of `globalThis` across multiple files
- Simplified code in various components by removing unnecessary imports and using standardized global object access
- Updated package.json files to optimize dependencies and reduce bundle size

These changes align with the goal of reducing Storybook's install size and dependency count, as outlined in issue #29038.

<!-- /greptile_comment -->